SWIG_AsCharPtrAndSize simplification and performance improvement
Only make a copy of the returned string if the passed in Python object
does not own the returned char* string. Callers of SWIG_AsCharPtrAndSize
should now ensure that the returned string is only used while the passed
in PyObject * string is valid. Performance improvement is for Python 3
input string handling.

The new SWIG_PyUnicode_AsUTF8AndSize function allows for this simplification.