Friend operator overloading fix for Python builtin
Don't generate calls to
  SWIGPY_BINARYFUNC_CLOSURE(_wrap___lshift__)
for the 'global' function wrappers for the friend functions.

Only occurred when a friend class is within a namespace due to the
renames in pyopers.swg, such as:
  %rename(__lshift__) *::operator<<;
The rename was intended for member functions but was also being attached
to friends that are within a namespace.