Improve ccomplextest.i's check for C complex support
Support for complex was added in C99, so we should check
__STDC_VERSION__ >= 199901L rather than 199001L.  The previous
check happened to work with a C90 compiler because that won't
define __STDC_VERSION__, but was wrong for a C94 compiler.

C11 specified that __STDC_NO_COMPLEX__ could be defined by
compilers without complex support, so check that isn't
defined too.