[ocaml] Strip out the value macro typedef hack
It's far from ideal that ocaml's C API uses a common term like "value"
as a typedef as it can clash with uses of "value" as an identifier in
the C/C++ API being wrapped, but that's not a problem that SWIG created
or really is able to solve.

Previously SWIG would try to rename "value" inside ocaml's headers
by defining it as a macro before including them and then undefining it
afterwards.  However this doesn't work for uses of "value" in macros
defined in the ocaml headers for use in code using the ocaml C API.
This lead to SWIG trying to have tweaked copies to replace all such
macros, which have inevitably diverged from those defined by ocaml;
also SWIG lacks replacements for more recently introduced macros.