Check whether the D compiler specified via --with-d2-compiler is functional
The D compiler can be specified by the user via the option
--with-d2-compiler of configure. The specified program will then be
used for compiling the D examples and unit tests (more precisely, in
Examples/Makefile.in).

If the user specifies a D compiler that does not exist in the system
or is not functional, then the D tests will fail. This commit
introduces the check for the specified program in configure.ac,
unsetting the D2COMPILER variable if necessary. This ensures that the
D tests will not be compiled if the user specifies a non-working D
compiler.