Strict Python stable abi conformance fix for PyUnicode_GetLength
PyUnicode_GetLength is officially documented as being in the stable abi
from python-3.7 onwards. It is is actually available in the Python
headers when compiling with Py_LIMITED_API set to 0x03030000 (python-3.3)
and later. This change ensures it is only used when Py_LIMITED_API is
set to 0x03070000 (python-3.7) and later for strict conformance.