Fix specialization without primary template
Due to a regression in 817c700 (Template partial specialization
improvements, 2023-01-30), specializations of templates without a
primary template definition were not handled correctly any more and,
notably, resulted in misparsing the injected name inside such
specializations.

Fix this by handling the nodes with template type "classforward" created
for the template declaration in the same way as those created for the
primary template definition.

Add a trivial test case which works (again) now but doesn't work without
the fix in this commit.

Closes #2934.