Fix running preproc_include test with JS+node more than once
At least when building in a separate build directory, building this test
second time fails due to a syntax error in the makefile created by
node-gyp which happens because it doesn't handle dependencies with
spaces in them correctly.

Fix this by simply avoiding using such dependencies when using node-gyp,
this is not a big loss as the goal here is to test SWIG preprocessor and
not JS backend itself, and the preprocessor is still tested with the
other languages.

Closes #2930.