Restore C# empty() containers method but warn when using it
In d8c21410e (Replace empty() method with IsEmpty C# property in STL
containers, 2023-11-10) the previously existing empty() method was made
private, making it impossible to use it by default and breaking existing
code that called it and worked fine with the previous SWIG versions.

Instead of breaking it completely, still define this method but
deprecate it using the standard Obsolete() attribute and a message
pointing to IsEmpty property.