Pass overloaded function itself to add_method() for clarity
It's unclear why did we pass the previous element of the overload set
when calling add_method() from dispatchFunction() but it seems that it
was done just to avoid passing anything at all if this function was not
overloaded, presumably because whatever add_method() does if the "node"
argument is not null was already done from elsewhere in this case.

Unfortunately, add_method() doesn't document when should this argument
be null and when it shouldn't be, so it's just guesswork, but at least
the test suite still passes after this change which makes the code a bit
more clear.

And now we use the properties of this node itself in add_method() when
the function is overloaded instead of, completely inexplicably, using
the properties of another overloaded function (which was probably mostly
fine in practice until now because most overloaded functions have the
same properties anyhow).