Also fix docstrings when using Python with -builtin option
Use the recently added find_overload_with_docstring() in builtin case
too.

Note that this relies on the parent commit as add_method() must be
called with the last function in the overload set, and not the last but
one, for find_overload_with_docstring() to work correctly.

Alternatively, we could change find_overload_with_docstring() to look at
all elements of the overload set, not just the preceding, but also the
following ones. However it seems better to keep things simpler and only
call it for the last one.