Use -fPIC on Linux instead of -fpic
On most platforms, these are identical.  However, on a small list of
platforms, the small -fpic will enforce a limit on GOT size.  However,
if -fPIC is used instead, then the GOT on these platforms will have no
limits.  The affected platforms and the respective GOT limits with -fpic
are:  8k on sparc, 28k on aarch64, and 32k on m68k and powerpc.

See: https://gcc.gnu.org/onlinedocs/gcc/Code-Gen-Options.html#index-fpic