Check C# xml documentation for all Doxygen tests
The C# compiler is now passed -doc for all Doxygen tests for wider
testing of the generated documentation.

The @see tag translation into <seealso> does not convert the parameter
list from C++ to C#, so for now, just remove the parameters.
May give warning CS0419 if the method is ambiguous, so this is also
suppressed in the test-suite.