Fix duplicate friend wrappers for friend declarations in nested classes.
We need to modify the scope that friends are declared in to the
namespace that of the outermost class is declared in. The friend is
not declared within a class, it must be a namespace. Add missing
implementations of the "prev_symtab" so that it is set in every
class and then use this for picking up the appropriate scope for
declaring the friend in.

Closes #2845