Remove Cygwin testing
I started off developing SWIG with Cygwin in 2003, but times have
changed and Cygwin is not particularly relevant now with WSL and
evolutions resulting in MinGW-64 and MSYS2 on Windows.
Additionally testing Cygwin doesn't offer much extra nowadays and I feel
comfortable dropping this support in favour of MinGW-64+MSYS2 done in
.github/workflows/windows.yml.

Part of the migration from appveyor to Github Actions.