Fix JS keyword rename test to actually test for renaming
Assigning to a property in JS always works, as it just creates a new
property if there is no existing one, and so can't be used for testing
for its existence.

Set the property value in C code and try accessing it from JS instead.

Also check that not only the property with the expected name exists, but
also that the property with the renamed name does not.

See #2943.