Drop bogus handling of SWIG_TOKEN_COMMENT in yylex
SWIG_TOKEN_COMMENT is a token value from scanner.c, but we're switching
on a parser token value here.

As the code currently is this case will never get taken since the parser
token value which gets allocated the same numeric value as
SWIG_TOKEN_COMMENT is TYPE_SIGNED, which isn't returned by yylook(), but
there's potential for compilation error if this token number was
allocated to a token which is another case in this switch, or for
incorrect behaviour if if this token number was allocated to a token
which yylook() can return which should take the default case in this
switch.

Comments are actually removed by the preprocessor, so we don't need to
concern ourselves with them here.