Drop support for raw type strings in interface files
This is described in CHANGES as a "Secret developer feature" and
a "SICK HACK".  It was only documented in developer documentation, and
had no test coverage.

It allowed specifying an SWIG internal syntax type string, e.g.

  `p.a(10).p.f(int, p.f(int).int)` foo(int, int (*x)(int));

However there seems to be no good reason to support this - it's
better to use the C/C++ syntax:

  (*(*foo(int, int (*)(int)))[10])(int, int (*)(int));

Fixes #2998
Closes #3034