[javascript] Fix v8 support and add CI coverage
Update the v8 support to work with recent v8 releases.

Fix bitrot in the testsuite and examples so they pass for v8.

Adjust configure probes so we can successfully probe for v8 without
assistance on Debian and Ubuntu (and probably other platforms; those
are just the two I've tested, and both use multiarch library
directories which weren't being allowed for).

Closes #3028
See #2537