Clean up callparms parser rules
This eliminates a case of right recursion - it's actually simpler
and more efficient to build the string up left to right.

We now use the str token type for callparms as we were only using
dtype's val member.

Empty callparms is now just NULL, which SWIG's Printf's %s format
handles as an empty strng.

Finally we now call Delete() to release strings we're finished with.

See #2884