Add regression test for old bug fix
We don't seem to have anything in the testsuite which exercises this
case.

There's some special handling of casts applied to string literals in the
parser which was committed as part of the fix for this, but removing it
doesn't seem to cause anything to fail (not even this testcase which
it reportedly fixed), and a wide character version seems to work without
any equivalent special handling.  Maybe other things have changed since
which fixed this case without needing special cast handling.  Deserves
further investigation.