Fix initializer lists used with -keyword or compactdefaultargs option
Remove casts to local variables that have initial values taken from the
default arguments being wrapped. This is for the default case, that is,
for all types, except references.

This fixes handling of parameters with default arguments that are initializer
lists by removing a cast to the initializer.

For wrapping parameter X x = {}, the generated code previously would have
contained:
  X arg2 = (X) {};
Now it is:
  X arg2 = {};

Closes #1851