[guile] Fix ports.i
This contains a workaround intended to allow it to work when
fdopen() isn't available by default.  That was apparently the case
on Solaris, but is also true on Linux when using an option which
turns off GNU extensions such as -std=c11.

However the workaround doesn't work properly because stdio.h has
already been included earlier in the generated file.  Moving the
workaround to the "begin" section fixes this.

Discovered while trying to update the GHA CI to not use ubuntu-20.04.
The update guile jobs were failing, so once merged that will provide
test coverage for this change.