Strip additional -std flags that autoconf now adds to CC and CXX
Starting from autoconf 2.70, autoconf macros AC_PROG_CC and AC_PROG_CXX
decide which c and c++ standards should be used by adding compiler flags
to the program name. We need these macros to provide the compiler name
only as it used to do prior to 2.69. This commit strips the -std= option
off the CC and CXX variables returned by these two autoconf macros.

Currently, in the year 2025, autoconf thinks that the c++ standard
to use should be c++11. If this version of autoconf is the one being
used in 10 years time, it will still think that c++11 is the version to
use, this is never going to be what we want and then the c++ standards
being mandated will depend on the version of autoconf being used. This
is never going to be satisfactory.

SWIG has its own mechanisms to control which standard to use and test
against. These are based around the autoconf macro AX_CXX_COMPILE_STDCXX
and Github Actions (.github/workflows/*.yml) and usage of Tools/testflags.py.

If CXX is 'g++ -std=gnu++11' and we then use $CXX -std=c++23, this
should give test c++23 so adding the flags isn't such a big deal, but is
quite confusing seeing two -std options in the build.

However, when we don't specify the flags, we expect g++ to use the
standard that it is built by default. So autoconf adding -std flags is
just confusing at best.