GHA update for ubuntu-20.04 EOL (#3133)
We lose Linux builds with GCC 7 and 8
We lose Guile 2.0 testing
We lose Python 3.5 and 3.6 testing
We need to adjust things to be able to still test python2 (because the package to install is now python2-dev instead of python-dev and /usr/bin/python is now always Python3)
We lose mzscheme/racket testing (we have already flagged that support would be dropped at this point unless somebody updated SWIG to work with Racket 8: see 

Support for Racket 8 #2830)
I've had to comment out Ruby 2.4, 2.5, 2.6, 2.7 and 3.0 for now as installing them with rvm fails with an unhelpful error message

The Ruby one needs addressing. We could probably use Ubuntu's ruby3.0 package for 3.0, but 2.x isn't available as a package for ubuntu-22.04. It seems Ruby 3.0 is only just over 4 years old so losing testing of all 2.x wouldn't be ideal (losing the oldest/older 2.x would be reasonable at this point).

Closes #3132