Use C++11 std::mutex with MSVC too
Check for _MSVC_LANG in addition to checking for __cplusplus as MSVC
compiler still doesn't define the latter by default, even if supports
not just C++11 but even much later standards.

This notably avoids including <windows.h> from this file which wreaks
havoc with the code after it due to the definitions of many common names
as macros in it.