Python DeprecationWarning fixes
DeprecationWarning fixes when using 'python -Walways' or if using one of the types being warned about:

  DeprecationWarning: builtin type SwigPyPacked has no __module__ attribute
  DeprecationWarning: builtin type SwigPyObject has no __module__ attribute
  DeprecationWarning: builtin type swigvarlink has no __module__ attribute

The warning was present if code was generated without using -builtin in 4.3.x.
The warning was only present if using the Python limited API in 4.2.1 and earlier.

These 3 types were not in any module at all. Now they are in the swig runtime module.
Docs to follow.

Issue #2881
Issue #3061
Issue #3160