Fix regression expanding templates in function parameter values
Only using the parameter value in the generated code, such as when using
kwargs. This is known to be fundamentally flawed as sometimes the
generated value is not accessible outside of the class it is defined in.

Fixes regression due to a037e2f2e21884eb282b15da1478508d1cce5f86.
The new fix now attempts to treat all value and name attributes as
a SwigType instead of an unparsed type in a String. However, it does
it more consistently instead of just for the "value" attribute in
add_parms.

Also subtle problem fix where cpatchlist was incorrectly being used
instead of patchlist.

Closes #3179