Add SWIG_PYTHON_OLD_BUFFER_PROTOCOL fragments for pybuffer typemaps
The pybuffer typemaps use a new fragment for issuing the helpful warning
to help clarify when the old Buffer Protocol can be used. This prevents
it being issued when the typemaps are not used and makes it clearer in
the generated code when the old Protocol Buffer is being used.

Define SWIG_NO_PYTHON_OLD_BUFFER_PROTOCOL to override the choice of when
to define SWIG_PYTHON_OLD_BUFFER_PROTOCOL.