Improve wchar_t handling
SWIG previously always treated wchar_t as a keyword, which it is in
C++ but not in C.  So for example `typedef unsigned short wchar_t;`
was rejected with a parse error even in C mode.

Also MSVC seems to allow such a typedef even in C++ mode.

SWIG now has a __WCHAR_TYPE__ keyword, and we typedef wchar_t to
that so the typedef above no longer causes a parse error (you will
still get a warning).  This is done in both C and C++ mode, so
C++ headers relying on MSVC's laxness can also now be parsed.

Fixes #3081