Modify Python StopIteration implementation to work across multiple modules
Use the Python C API to set the StopIteration as soon as possible in
and then use NULL return or PyErr_Occurred() to detect handle the
StopIteration once raised while still in C code. This removes the need
to thrown the stop_iteration C++ exception, which cannot be caught
across different shared objects on some operating
systems/compiler configurations.

Closes #3189