[ci] Consistently suppress progress from checkout
The progress messages bloat the log and make it harder to search
for the relevant part.

The only case where they're even marginally useful is when checkout
fails, as you can at least see how far it gets, but even then they
don't really seem useful.

Raised as a side issue in #3191